feat: merge router
