+++ routes/timetable.py
@@ -1,18 +1,27 @@
-from fastapi import APIRouter, Depends, HTTPException
+from fastapi import APIRouter, Depends, HTTPException, Query
+from fastapi.params import Param
 
 from core.auth.client import get_user
 from core.neis.time_table import TimeTable
 from depends.require_auth import RequireAuth
 
+from pydantic import BaseModel
+
 router = APIRouter(prefix="/timetable")
 
+class TimetableDTO(BaseModel):
+    department: str
+    grade: int
+    class_num: int
 
 @router.get("/")
-async def get_timetable(userid: int = Depends(RequireAuth)):
+async def teacher_timetable(
+    department: str = Query(None, description="Department name"),
+    grade: int = Query(None, description="Grade of the class"),
+    classroom: int = Query(None, description="Classroom number"),
+    userid: int = Depends(RequireAuth)
+    ):
     user = await get_user(userid)
-
-    if not user:
-        raise HTTPException(status_code=404, detail="USER_NOT_FOUND")
 
     if (
         not user.verification.department
@@ -20,12 +29,26 @@ async def get_timetable(userid: int = Depends(RequireAuth)):
         or not user.verification.classroom
     ):
         raise HTTPException(status_code=403, detail="USER_NOT_VERIFIED")
-
-    timetable = await TimeTable.get(
-        user.verification.department,
-        user.verification.grade,
-        user.verification.classroom,
-    )
+    
+    if (
+        not department
+        or not grade
+        or not classroom
+    ):
+        timetable = await TimeTable.get(
+            user.verification.department,
+            user.verification.grade,
+            user.verification.classroom,
+        )
+    else:
+        timetable = await TimeTable.get(
+            department,
+            grade,
+            classroom,
+        ) 
+        
+    if not timetable:
+        raise HTTPException(status_code=404, detail="NOT_FOUND")
 
     return {
         "message": "SUCCESS",
